requirements.txt
asgiref==3.10.0
Django==5.2.8
djangorestframework==3.16.1
pydub==0.25.1
sqlparse==0.5.3
typing_extensions==4.15.0
